Move rules and certificates logic to Interval
